Protect nested attribute bags and trim dead API surface
Finish the post-spread Omit pattern across Nav, Toggle, CodeExample,
FileDetails and DraggablePane so shared bags cannot clobber owned attrs
(href, type, bind:checked, menu id). Drop dialog_style and unexport
toast/utils internals unused outside their modules; fold duplicate
clipboard/toast tests and unit-test the previously untested utils.
